refactor: move binary eval to evaluation module, batch inference, save run environment
Move binary metrics pipeline (compute_p_toxic, run_binary_evaluation,
build_eval_loader) from training/orchestrator.py to evaluation/binary.py.
CLI eval_binary reduced from 60-line reimplementation to thin wrapper.

Batch run_inference() — reads all embeddings into a tensor, runs batched
forward passes (512 per batch) instead of one-by-one protein processing.

Save run_environment.json during training with git commit, torch version,
Python version, device, platform, and timestamp for reproducibility.
