+++ src/toxfam/model/inference.py
@@ -88,27 +88,38 @@ def run_inference(
     is_multi_input = model_config.architecture == "MultiInputMLP"
     tax_dim = model_config.tax_dim if is_multi_input else None
 
-    preds = []
-    confs = []
-    with h5py.File(h5_path, "r") as f:
-        for ident in df["identifier"]:
-            emb = torch.tensor(f[ident][:]).unsqueeze(0).to(device)
+    identifiers = df["identifier"].tolist()
 
-            with torch.no_grad():
-                if is_multi_input:
-                    dummy_tax = torch.zeros(1, tax_dim).to(device)
-                    logits = model(emb, dummy_tax)
-                else:
-                    logits = model(emb)
-                probs = torch.softmax(logits, dim=1)
-                conf, pred_idx = probs.max(dim=1)
-            preds.append(idx_to_label.get(pred_idx.item(), "other"))
-            confs.append(conf.item())
+    # Read all embeddings into a single tensor
+    with h5py.File(h5_path, "r") as f:
+        embeddings = torch.stack(
+            [torch.tensor(f[ident][:], dtype=torch.float32) for ident in identifiers]
+        )
+
+    # Batched forward pass
+    all_preds: list[str] = []
+    all_confs: list[float] = []
+    batch_size = 512
+
+    with torch.no_grad():
+        for i in range(0, len(embeddings), batch_size):
+            batch = embeddings[i : i + batch_size].to(device)
+            if is_multi_input:
+                dummy_tax = torch.zeros(batch.shape[0], tax_dim, device=device)
+                logits = model(batch, dummy_tax)
+            else:
+                logits = model(batch)
+            probs = torch.softmax(logits, dim=1)
+            confs, pred_idxs = probs.max(dim=1)
+            all_preds.extend(
+                idx_to_label.get(idx.item(), "other") for idx in pred_idxs
+            )
+            all_confs.extend(confs.cpu().tolist())
 
     return pd.DataFrame(
         {
-            "identifier": df["identifier"].values,
-            "predicted_label": preds,
-            "confidence": confs,
+            "identifier": identifiers,
+            "predicted_label": all_preds,
+            "confidence": all_confs,
         }
     )
