+++ src/toxfam/training/orchestrator.py
@@ -3,10 +3,8 @@ from __future__ import annotations
 import json
 from pathlib import Path
 
-import numpy as np
 import pandas as pd
 import torch
-import torch.nn.functional as F
 import yaml
 from rich.console import Console
 from torch.utils.data import DataLoader
@@ -25,137 +23,10 @@ from toxfam.training.strategies import (
     run_combined_strategy,
     run_standard_strategy,
 )
-from toxfam.training.trainer import forward_model, get_class_weights, get_device, set_seed
-from toxfam.visualization.analysis import (
-    analyze_label_distribution_for_split,
-    plot_binary_pr,
-    plot_binary_roc,
-)
+from toxfam.training.trainer import get_class_weights, get_device, set_seed
+from toxfam.visualization.analysis import analyze_label_distribution_for_split
 
 console = Console()
-
-
-def compute_binary_labels(df: pd.DataFrame, label_col: str = "Protein families") -> np.ndarray:
-    """Convert family labels to binary: 1 = toxic, 0 = nontoxin."""
-    from toxfam.evaluation.metrics import to_binary_class
-
-    return (df[label_col].apply(to_binary_class) == "toxin").astype(int).values
-
-
-def compute_p_toxic(
-    model, dataset_df, config, label_encoder, label_col="Protein families"
-) -> np.ndarray:
-    """Compute P(toxic) for each sample by summing toxic-class probabilities."""
-    from toxfam.evaluation.metrics import NONTOXIN_LABELS
-
-    ds, selector = build_eval_loader(dataset_df, config, label_encoder, label_col)
-
-    device = get_device()
-    model = model.to(device)
-    model.eval()
-
-    all_probs = []
-    with torch.no_grad():
-        for features, _ in selector:
-            outputs = forward_model(model, features, device)
-            probs = F.softmax(outputs, dim=1).cpu().numpy()
-            all_probs.append(probs)
-
-    all_probs = np.concatenate(all_probs, axis=0)
-    ds.close()
-
-    # Sum probabilities of all nontoxin classes
-    nontox_indices = [
-        i for i, cls in enumerate(label_encoder.classes_)
-        if cls.lower() in NONTOXIN_LABELS
-    ]
-    p_nontox = all_probs[:, nontox_indices].sum(axis=1)
-    return 1.0 - p_nontox
-
-
-def run_binary_metrics_pipeline(
-    model, train_ds, val_df, test_df, config, out_root, label_col="Protein families"
-):
-    """Full binary metrics pipeline: threshold optimization on val, evaluate on test."""
-    from toxfam.evaluation.metrics import (
-        calculate_binary_metrics_with_scores,
-        find_optimal_threshold,
-    )
-
-    console.print("\n[bold]Running Binary Metrics Pipeline...[/bold]")
-
-    # Val set
-    val_y_true = compute_binary_labels(val_df, label_col)
-    val_p_toxic = compute_p_toxic(model, val_df, config, train_ds.le, label_col)
-
-    # Threshold optimization on val
-    thresh_result = find_optimal_threshold(val_y_true, val_p_toxic, method="youden")
-    opt_threshold = thresh_result["optimal_threshold"]
-    console.print(f"  Optimized threshold (Youden's J): {opt_threshold:.4f}")
-
-    # Test set
-    test_y_true = compute_binary_labels(test_df, label_col)
-    test_p_toxic = compute_p_toxic(model, test_df, config, train_ds.le, label_col)
-
-    # Test with default threshold
-    test_default = calculate_binary_metrics_with_scores(test_y_true, test_p_toxic, threshold=0.5)
-    console.print(
-        f"  Test (t=0.5): ROC-AUC={test_default['roc_auc']:.4f}, "
-        f"PR-AUC={test_default['pr_auc']:.4f}, MCC={test_default['mcc']:.4f}"
-    )
-
-    # Test with optimized threshold
-    test_opt = calculate_binary_metrics_with_scores(test_y_true, test_p_toxic, threshold=opt_threshold)
-    console.print(
-        f"  Test (t={opt_threshold:.3f}): ROC-AUC={test_opt['roc_auc']:.4f}, "
-        f"PR-AUC={test_opt['pr_auc']:.4f}, MCC={test_opt['mcc']:.4f}"
-    )
-
-    # Save metrics (curve data already serializable lists)
-    metrics_dir = out_root / "metrics"
-    metrics_dir.mkdir(exist_ok=True)
-    # Filter out large curve arrays for the summary JSON
-    _curve_keys = {"fpr", "tpr", "precision_curve", "recall_curve", "roc_thresholds", "pr_thresholds"}
-    binary_results = {
-        "optimized_threshold": opt_threshold,
-        "test_default": {k: v for k, v in test_default.items() if k not in _curve_keys},
-        "test_optimized": {k: v for k, v in test_opt.items() if k not in _curve_keys},
-    }
-    (metrics_dir / "binary_metrics.json").write_text(json.dumps(binary_results, indent=4))
-
-    # Plots
-    plots_dir = out_root / "plots"
-    plots_dir.mkdir(exist_ok=True)
-    plot_binary_roc(
-        test_default["fpr"], test_default["tpr"], test_default["roc_auc"],
-        plots_dir / "binary_roc.png",
-    )
-    plot_binary_pr(
-        test_default["precision_curve"], test_default["recall_curve"],
-        test_default["pr_auc"], plots_dir / "binary_pr.png",
-    )
-
-
-def build_eval_loader(
-    dataset_df, config: TrainConfig, label_encoder, label_col="Protein families",
-):
-    """Build a ToxDataset + DataLoader + DataSelector for evaluation.
-
-    Returns (dataset, selector) — caller must call dataset.close() when done.
-    """
-    ds = ToxDataset(
-        dataset_df,
-        [str(p) for p in config.h5_paths],
-        label_encoder=label_encoder,
-        is_train=False,
-        label_col=label_col,
-        tax_h5_path=str(config.tax_h5_path) if config.tax_h5_path else None,
-    )
-    loader = DataLoader(ds, batch_size=config.batch_size, shuffle=False)
-    selector = DataSelector(
-        loader, "both" if config.training_strategy == "combined" else "emb_only",
-    )
-    return ds, selector
 
 
 def run_training(config: TrainConfig) -> None:
@@ -194,6 +65,23 @@ def run_training(config: TrainConfig) -> None:
     predictions_dir = out_root / "predictions"
     for d in (models_dir, plots_dir, metrics_dir, predictions_dir):
         d.mkdir(exist_ok=True)
+
+    # Save run environment for reproducibility
+    import platform
+    import sys
+    from datetime import datetime
+
+    from toxfam.evaluation.runner import git_commit_short
+
+    env_info = {
+        "timestamp": datetime.now().isoformat(),
+        "git_commit": git_commit_short(),
+        "python_version": sys.version,
+        "torch_version": torch.__version__,
+        "device": str(device),
+        "platform": platform.platform(),
+    }
+    (out_root / "run_environment.json").write_text(json.dumps(env_info, indent=2))
 
     # 1. Load Data
     console.print("Loading data...")
@@ -382,8 +270,10 @@ def run_training(config: TrainConfig) -> None:
     # For binary strategy, the model directly outputs binary classes,
     # so we pass the effective label col. For other strategies, we derive
     # binary from the original family labels.
-    run_binary_metrics_pipeline(
-        scaled_model, train_ds, val_df, test_df, config, out_root, effective_label_col
+    from toxfam.evaluation.binary import run_binary_evaluation
+
+    run_binary_evaluation(
+        scaled_model, train_ds.le, val_df, test_df, config, out_root, effective_label_col
     )
 
     train_ds.close()
